Added cloud storage interface
